update "GM_getUrl()" to support execution in DOMContentLoaded event
The fundamental issue is that the DOMContentLoaded event
fires before WebViewClient.onPageFinished(),
and the code that tracks the current URL is updated in the latter.

This code is ridiculously convoluted and complex,
and it is meant to deal with complications that originate in WebView;
specifically, that extraneous onPageStarted() events occur for a URL
after onPageFinished() has already been called,
and after onPageStarted() has been called for a new URL..
it's basically a sh*t show.

This update hooks into the existing methodology
that tracks onPageStarted() URLs,
but doesn't finalize them until onPageFinished().
When a userscript executes in the DOMContentLoaded event,
and immediately calls GM_getUrl(),
the current URL is not yet available,
but a different variable holds the most recent onPageStarted() URL.
This not-yet-finalized URL is now the value that is returned.